bas-data: IMemory.read() returns int instead of byte.
- compatible with I/O interfaces.
- Returns -1 for END.
- no need to cast to int.

+++ bas/layer-3/bas-data/src/main/java/net/bodz/bas/data/mem/IMemory.java
@@ -5,7 +5,7 @@ package net.bodz.bas.data.mem;
  */
 public interface IMemory {
 
-    byte read(int addr)
+    int read(int addr)
             throws MemoryAccessException;
 
     void write(int addr, byte value)
